Adding remaining integer conversions

+++ bietlejuice/db/3.dw/growth/prod/queries/supply/leads/prefix_city.sql
@@ -55,8 +55,8 @@ all_dates_last_month as (
   	on f.sk_property = dpr.sk_property
 	left join dim_region dr
 		on dpr.regiao_id = dr.id
-	where f.sk_lead_date between to_char(add_months(date_trunc('month', current_date), -1), 'YYYYMMDD')
-    and to_char(add_months(current_date - 1, -1), 'YYYYMMDD')
+	where f.sk_lead_date between to_char(add_months(date_trunc('month', current_date), -1), 'YYYYMMDD')::integer
+    and to_char(add_months(current_date - 1, -1), 'YYYYMMDD')::integer
  	group by 4, 1, 2
   order by 4, 1, 2
 ),
@@ -71,8 +71,8 @@ all_dates_last_year as (
   	on f.sk_property = dpr.sk_property
   left join dim_region dr
   	on dpr.regiao_id = dr.id
-  where f.sk_lead_date between to_char(add_months(date_trunc('year', current_date), -12), 'YYYYMMDD')
-    and to_char(add_months(current_date - 1, -12), 'YYYYMMDD')
+  where f.sk_lead_date between to_char(add_months(date_trunc('year', current_date), -12), 'YYYYMMDD')::integer
+    and to_char(add_months(current_date - 1, -12), 'YYYYMMDD')::integer
   group by 3, 1
 	order by 3, 1
 ),
